refactor: apply feedback and modularize mostpopular reranking logic
